Tidy up. #21
